Added clustering and vgg stuff

+++ model/network.py
@@ -311,7 +311,7 @@ class Converter:
 		if self.config.img_shape[-1] == 1:
 			x = Lambda(lambda t: tf.tile(t, multiples=(1, 1, 1, 3)))(img)
 		elif self.config.img_shape[-1] == 2:
-			paddings = K.constant([[0, 0], [0, 0], [0, 0], [0, 1]])
+			paddings = K.constant([[0, 0], [0, 0], [0, 0], [0, 1]], dtype=tf.int32)
 			# channel_to_add = np.zeros(shape=(self.config.img_shape[0], self.config.img_shape[1], 1), dtype=np.float32)
 			x = Lambda(lambda t: tf.pad(t, paddings=paddings, constant_values=0))(img)
 			# x = K.map_fn(lambda t: tf.concat((t, [channel_to_add]), axis=-1), img)
